Qiro coder Telegram bot try/requirements.txt
alembic
aiosqlite
asyncpg
fastapi
httpx
psycopg2-binary
prometheus-client
pydantic
pytest
pytest-asyncio
python-dotenv
sqlalchemy
telethon
uvicorn
selenium
webdriver-manager
beautifulsoup4
lxml
requests